+++ lib/src/places/services/api/places_api.g.dart
@@ -1,0 +1,82 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'places_api.dart';
+
+// **************************************************************************
+// RetrofitGenerator
+// **************************************************************************
+
+class _PlacesApi implements PlacesApi {
+  _PlacesApi(this._dio, {this.baseUrl}) {
+    baseUrl ??= 'https://stoplight.io/mocks/softech/wtgt/75539';
+  }
+
+  final Dio _dio;
+
+  String? baseUrl;
+
+  @override
+  Future<PlaceListResponse> fetchPlaces({search, lat, lon, isFavourite}) async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{
+      r'search': search,
+      r'lat': lat,
+      r'lon': lon,
+      r'fav': isFavourite
+    };
+    queryParameters.removeWhere((k, v) => v == null);
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<PlaceListResponse>(
+            Options(method: 'GET', headers: _headers, extra: _extra)
+                .compose(_dio.options, '/places',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = PlaceListResponse.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<Place> fetchPlace(id) async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<Place>(
+            Options(method: 'GET', headers: _headers, extra: _extra)
+                .compose(_dio.options, '/places/${id}',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = Place.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<void> setFavourite(id, body) async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = body;
+    await _dio.fetch<void>(_setStreamType<void>(
+        Options(method: 'PUT', headers: _headers, extra: _extra)
+            .compose(_dio.options, '/places/${id}/favorite',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    return null;
+  }
+
+  RequestOptions _setStreamType<T>(RequestOptions requestOptions) {
+    if (T != dynamic &&
+        !(requestOptions.responseType == ResponseType.bytes ||
+            requestOptions.responseType == ResponseType.stream)) {
+      if (T == String) {
+        requestOptions.responseType = ResponseType.plain;
+      } else {
+        requestOptions.responseType = ResponseType.json;
+      }
+    }
+    return requestOptions;
+  }
+}