+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - feat: MainScreen + PlacesScreen
 - feat: OnboardStorage
 - feat: profile screen
+- feat: profile Api
+- feat: places api
+- feat: profile slivers
 
 ## 0.3.0